third_party/blink/web_tests/accessibility/css-generated-content-expected.txt

Tests that whitespace is properly collapsed in CSS generate content, and that the concatenation of all of the strings from inline text boxes matches the string value of the text object.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS text is concatenatedInlineText
PASS text is concatenatedInlineText
PASS text is concatenatedInlineText
PASS successfullyParsed is true

TEST COMPLETE


six seven eight nine ten 
